Lot of optimisation (Frustum + Greedy Meshing) Collisions added (#Thanks JBox2D) Directions fixed Created a much bigger level Tested some textures Added Water and Poison floor (W.I.P)

+++ app/src/main/java/com/gp5/projettutore/game/render/GUI/FontRenderer.java
@@ -34,8 +34,8 @@ public class FontRenderer
 			}
 			else
 			{
-				float tx = ((int) (character % GRID_SIZE)) * CELL_SIZE;
-				float ty = ((int) (character / GRID_SIZE)) * CELL_SIZE;
+				float tx = character % GRID_SIZE * CELL_SIZE;
+				float ty = character / GRID_SIZE * CELL_SIZE;
 
                 quadList.add(new Textured2DQuad(Textures.fontTexture, 16, 16, tx, ty, tx + 0.0625F, ty + 0.0625F));
 			}
